radix_heap done

+++ lista3/lib.hpp
@@ -2,10 +2,11 @@
 
 #include <list>
 #include <cstddef>
+#include <cstdint>
 #include <vector>
 
-typedef std::vector<std::pair<size_t, std::list<size_t>::iterator> > distances;
-typedef std::pair<size_t, size_t> pair;
+typedef std::vector<std::pair<int32_t, std::list<int32_t>::iterator> > distances;
+typedef std::pair<int32_t, int32_t> pair;
 
 class Graph
 {
@@ -16,11 +17,11 @@ class Graph
     public:
         Graph(const size_t node_quantity, const size_t max_weight);
 
-        auto add_edge(const size_t vert1, const size_t vert2, const size_t weight) -> void;
+        auto add_edge(const int32_t vert1, const int32_t vert2, const size_t weight) -> void;
         auto dijkstra_classic_ss(const size_t src) -> distances;
-        auto dijkstra_classic_p2p(const size_t start, const size_t goal) -> size_t;
+        auto dijkstra_classic_p2p(const size_t start, const size_t goal) -> int32_t;
         auto dial_ss(const size_t src) -> distances;
-        auto dial_p2p(const size_t start, const size_t goal) -> size_t;
+        auto dial_p2p(const size_t start, const size_t goal) -> int32_t;
         auto radix_heap_ss(const size_t src) -> std::vector<int>;
-        auto radix_heap_p2p(const size_t start, const size_t goal) -> size_t;
+        auto radix_heap_p2p(const size_t start, const size_t goal) -> int32_t;
 };
